+++ R/ggeffect.R
@@ -1,6 +1,5 @@
 #' @rdname ggpredict
 #'
-#' @importFrom purrr map2
 #' @importFrom stats na.omit
 #' @importFrom sjlabelled as_numeric
 #' @importFrom insight find_predictors link_inverse print_color
@@ -17,9 +16,9 @@ ggeffect <- function(model, terms, ci.lvl = .95, ...) {
     terms <- all.vars(terms)
   }
 
-  if (inherits(model, "list"))
+  if (inherits(model, "list")  && !inherits(model, c("bamlss", "maxLik"))) {
     res <- lapply(model, function(.x) ggeffect_helper(.x, terms, ci.lvl, ...))
-  else {
+  } else {
     if (missing(terms) || is.null(terms)) {
       predictors <- insight::find_predictors(model, effects = "fixed", component = "conditional", flatten = TRUE)
       res <- lapply(
@@ -130,7 +129,7 @@ ggeffect_helper <- function(model, terms, ci.lvl, ...) {
     eff.logits <- as.data.frame(eff$logit, stringsAsFactors = FALSE)
     tmp <- cbind(eff$x, eff.logits)
     ft <- (ncol(tmp) - ncol(eff.logits) + 1):ncol(tmp)
-    tmp <- .gather(tmp, "response.level", "predicted", colnames(tmp)[ft])
+    tmp <- .gather(tmp, names_to = "response.level", values_to = "predicted", colnames(tmp)[ft])
 
     fx.term <- eff$term
 
@@ -147,7 +146,7 @@ ggeffect_helper <- function(model, terms, ci.lvl, ...) {
     # compute CI manually and then also fix column names.
 
     eff.se.logits <- as.data.frame(eff$se.logit)
-    tmp2 <- .gather(eff.se.logits, "response.level", "se", colnames(eff.se.logits))
+    tmp2 <- .gather(eff.se.logits, names_to = "response.level", values_to = "se", colnames(eff.se.logits))
     tmp2$conf.low <- tmp$predicted - stats::qnorm(ci) * tmp2$se
     tmp2$conf.high <- tmp$predicted + stats::qnorm(ci) * tmp2$se
     tmp2$std.error <- tmp2$se