+++ R/print.R
@@ -1,5 +1,3 @@
-#' @importFrom purrr flatten_df
-#' @importFrom sjmisc round_num add_variables
 #' @importFrom stats quantile
 #' @importFrom sjlabelled as_label get_labels
 #' @export
@@ -52,7 +50,7 @@ print.ggeffects <- function(x, n = 10, digits = 2, x.lab = FALSE, ...) {
   if (!is.null(a1) && !is.null(a2) && a1 == "coxph" && !(a2 == "Risk Score"))
     terms <- c("time", terms)
 
-  x <- sjmisc::round_num(x, digits = digits)
+  x <- .round_numeric(x, digits = digits)
 
   # if we have groups, show n rows per group
 
@@ -250,17 +248,14 @@ print.ggeffects <- function(x, n = 10, digits = 2, x.lab = FALSE, ...) {
 
 
 
-#' @importFrom insight format_table
+#' @importFrom insight format_table format_ci
 .print_block <- function(i, n, digits, ci.lvl, ...) {
   i <- i[setdiff(colnames(i), c("group", "facet", "panel", "response.level", ".nest"))]
   # print.data.frame(, ..., row.names = FALSE, quote = FALSE)
   dd <- i[.get_sample_rows(i, n), ]
 
   if ("conf.low" %in% colnames(dd) && "conf.high" %in% colnames(dd)) {
-    max_len_low <- max(unlist(lapply(stats::na.omit(round(dd$conf.low, digits)), function(.i) nchar(as.character(.i)))))
-    max_len_high <- max(unlist(lapply(stats::na.omit(round(dd$conf.high,digits)), function(.i) nchar(as.character(.i)))))
-
-    dd$CI <- .format_confint(dd$conf.low, dd$conf.high, digits = digits, width_low = max_len_low, width_high = max_len_high)
+    dd$CI <- insight::format_ci(dd$conf.low, dd$conf.high, digits = digits, width = "auto")
     dd$CI <- gsub("95% CI ", "", dd$CI, fixed = TRUE)
 
     if (is.null(ci.lvl)) ci.lvl <- .95
@@ -277,36 +272,4 @@ print.ggeffects <- function(x, n = 10, digits = 2, x.lab = FALSE, ...) {
   colnames(dd)[which(colnames(dd) == "predicted")] <- "Predicted"
   cat(insight::format_table(dd, digits = digits, protect_integers = TRUE))
   # print.data.frame(dd, ..., quote = FALSE, row.names = FALSE)
-}
-
-
-
-
-.format_confint <- function(CI_low, CI_high, ci = 0.95, digits = 2, brackets = TRUE, width = NULL, width_low = width, width_high = width) {
-  if (!is.null(ci)) {
-    ifelse(is.na(CI_low) & is.na(CI_high), "", paste0(ci * 100, "% CI ", .format_ci(
-      CI_low,
-      CI_high,
-      digits = digits,
-      brackets = brackets,
-      width_low = width_low,
-      width_high = width_high
-    )))
-  } else {
-    ifelse(is.na(CI_low) & is.na(CI_high), "", .format_ci(
-      CI_low,
-      CI_high,
-      digits = digits,
-      brackets = brackets,
-      width_low = width_low,
-      width_high = width_high
-    ))
-  }
-}
-
-
-
-#' @importFrom insight format_value
-.format_ci <- function(CI_low, CI_high, digits = 2, brackets = TRUE, width_low = NULL, width_high = NULL) {
-  paste0(ifelse(isTRUE(brackets), "[", ""), insight::format_value(CI_low, digits = digits, missing = "missing", width = width_low), ", ", insight::format_value(CI_high, digits = digits, missing = "missing", width = width_high), ifelse(isTRUE(brackets), "]", ""))
 }