+++ R/utils.R
@@ -1,9 +1,3 @@
-#' @importFrom magrittr %>%
-#' @export
-magrittr::`%>%`
-
-
-
 data_frame <- function(...) {
   x <- data.frame(..., stringsAsFactors = FALSE)
   rownames(x) <- NULL
@@ -48,7 +42,7 @@ data_frame <- function(...) {
 #' @importFrom sjlabelled as_label as_numeric
 .get_raw_data <- function(model, mf, terms) {
   # for matrix variables, don't return raw data
-  if (any(purrr::map_lgl(mf, is.matrix)) && !inherits(model, c("coxph", "coxme")))
+  if (any(sapply(mf, is.matrix)) && !inherits(model, c("coxph", "coxme")))
     return(NULL)
 
   if (!all(insight::find_response(model, combine = FALSE) %in% colnames(mf)))